Refactor course and instructor services; add enrollment
- Updated `CourseService` to rename `Page` to `PageNumber` and added logic to prevent deletion of courses with active enrollments.
- Introduced `GetTopRatedInstructors` method in `InstructorService` for retrieving instructors by rating with pagination.
- Modified `CreateInstructor` method for clearer `ImageUrl` assignment.
- Changed `Time` property type from `double` to `string` in `CourseLectureDto`, `CourseLectureToReturnDto`, and `CourseLecture` classes.
- Enhanced `ApplicationUser` and `Course` classes to include `CourseEnrollment` lists for tracking enrollments.
- Created `EnrollmentService`, `EnrollmentDto`, and `CourseEnrollment` classes to manage enrollment logic and data.
- Added `IEnrollmentService` interface and `CourseEnrollmentConfiguration` for entity configuration.
- Created a migration to add the `CourseEnrollment` table and update the `CourseLecture` table.
- Implemented `EnrollmentsController` for handling enrollment-related HTTP requests.
- Updated `Program.cs` for new service registrations and necessary namespaces.

+++ Byway.Application/Services/CourseService.cs
@@ -26,7 +26,7 @@ public class CourseService : ICourseService
     public async Task<PaginationModel<List<CourseListToReturnDto>>> GetAllCoursesAsync(CourseQueryModel courseQueryModel)
     {
         var courseRepository = _unitOfWork.GetRepository<Course>();
-        var pageNumber = courseQueryModel.Page;
+        var pageNumber = courseQueryModel.PageNumber;
         var pageSize = courseQueryModel.PageSize;
         var totalHours = courseQueryModel.TotalHours;
         var name = courseQueryModel.Name;
@@ -158,6 +158,10 @@ public class CourseService : ICourseService
         var existingCourse = await courseRepository.GetByIdAsync(id)
             ?? throw new NotFoundException("Course not found");
 
+        var enrollmentInstructor = _unitOfWork.GetRepository<CourseEnrollment>();
+        var enrollments = await enrollmentInstructor.GetAllAsync(q => q.Where(e => e.CourseId == id));
+        if(enrollments.Any())
+            throw new BadRequestException("Cannot delete course with active enrollments");
         courseRepository.Delete(existingCourse);
 
         if(existingCourse.ImageUrl is not null)
